Move secondary panel preferences to being saved on a per-replay basis
