+++ src/ui/actions/layout.ts
@@ -2,8 +2,7 @@ import { RecordingId } from "@recordreplay/protocol";
 import { Action } from "redux";
 import { getShowCommandPalette } from "ui/reducers/layout";
 import { dismissLocalNag, isLocalNagDismissed, LocalNag } from "ui/setup/prefs";
-import { ViewMode } from "ui/state/layout";
-import { PrimaryPanelName } from "ui/state/app";
+import { ViewMode, PrimaryPanelName, SecondaryPanelName } from "ui/state/layout";
 import { asyncStore } from "ui/utils/prefs";
 import { trackEvent } from "ui/utils/telemetry";
 import { UIThunkAction } from ".";
@@ -12,15 +11,17 @@ type SetSelectedPrimaryPanelAction = Action<"set_selected_primary_panel"> & {
   panel: PrimaryPanelName;
 };
 type SetShowCommandPalette = Action<"set_show_command_palette"> & { value: boolean };
-export type SetShowEditorAction = Action<"set_show_editor"> & {
+type SetShowEditorAction = Action<"set_show_editor"> & {
   showEditor: boolean;
 };
-export type SetShowVideoPanelAction = Action<"set_show_video_panel"> & {
+type SetShowVideoPanelAction = Action<"set_show_video_panel"> & {
   showVideoPanel: boolean;
 };
 type SetViewMode = Action<"set_view_mode"> & { viewMode: ViewMode };
+export type SetSelectedPanelAction = Action<"set_selected_panel"> & { panel: SecondaryPanelName };
 
 export type LayoutAction =
+  | SetSelectedPanelAction
   | SetSelectedPrimaryPanelAction
   | SetShowCommandPalette
   | SetShowEditorAction
@@ -58,6 +59,10 @@ export function setShowVideoPanel(showVideoPanel: boolean): SetShowVideoPanelAct
 
 export function setShowEditor(showEditor: boolean): SetShowEditorAction {
   return { type: "set_show_editor", showEditor };
+}
+
+export function setSelectedPanel(panel: SecondaryPanelName): SetSelectedPanelAction {
+  return { type: "set_selected_panel", panel };
 }
 
 export function setSelectedPrimaryPanel(panel: PrimaryPanelName): SetSelectedPrimaryPanelAction {
